pyproject.toml
[tool.poetry]
name = "women-protection"
version = "0.1.0"
description = ""
authors = ["Tanish2002 <tanishkhare@gmail.com>"]
readme = "README.md"
packages = [{include = "women_protection"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
bcrypt = "^4.0.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
psycopg2-binary = "^2.9.7"
pinecone-client = "^2.2.2"
openai = "^0.28.0"
langchain = "^0.0.288"
tiktoken = "^0.5.1"
unstructured = "^0.10.14"
django = "^5.0"
python-socketio = {extras = ["client"], version = "^5.10.0"}

[tool.pyright]
include = ["dependencies", "handlers", "models", "controllers", "configuration"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
